Vehicle improvements

+++ zword/src/util.c
@@ -157,7 +157,7 @@ int max(int a, int b) {
 }
 
 float mod(float x, float y) {
-    return fmod(fmod(x, y) + y, y);
+    return fmodf(fmodf(x, y) + y, y);
 }
 
 float cross(sfVector2f v, sfVector2f u) {
